Move generator test yaml files and rename test classes
Rellocate generator test yaml files to a sub folder
Rename generator test files to match the test subject
