+++ test/Atc.Rest.ApiGenerator.Tests/SyntaxGenerators/Api/XUnitTestData/ContractInterface/getSimpleInterface.yaml.verified.cs
