+++ test/Atc.Rest.ApiGenerator.Tests/SyntaxGenerators/Api/XUnitTestData/ContractParameter/getWithStringPath.yaml.verified.cs
