+++ test/Atc.Rest.ApiGenerator.Tests/SyntaxGenerators/Api/XUnitTestData/ContractParameter/getWithStringQuery.yaml.verified.cs
