+++ test/Atc.Rest.ApiGenerator.Tests/SyntaxGenerators/Api/XUnitTestData/EndpointControllers/getItem.yaml.verified.cs
